requirements.txt
beautifulsoup4>=4.12.3
click>=8.1.7
feedparser>=6.0.11
openai>=1.30.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
requests>=2.31.0
